vlab-api/vlab-operator-master-2/requirement.txt
async-timeout==3.0.1
attrs==21.2.0
cachetools==4.2.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
ecdsa==0.17.0
google-auth==2.3.2
idna==3.3
inflection==0.5.1
Jinja2==3.0.2
kubernetes==19.15.0
MarkupSafe==2.0.1
multidict==5.2.0
oauthlib==3.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
python-jose==3.3.0
python-keycloak==0.26.1
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
typing-extensions==3.10.0.2
urllib3==1.26.7
websocket-client==1.2.1
yarl==1.7.0
